Add Kubernetes cluster
This commit adds configured Kubernetes cluster to
the prompt. This is very useful if working with
several clusters simultaneously.
Many times we configure one cluster with
manifests intended for another cluster, so now
mistakes of this kind will be easier to avoid.

The addition relies on `kubectl` and `yq` installed
on the system. If not, no error will occur.